+++ integrations/github/src/types.ts
@@ -79,3 +79,28 @@ export type GithubConfigureState = Omit<
     withCustomTemplate?: boolean;
     commitMessagePreview?: string;
 };
+
+export type IntegrationTaskType = 'import:spaces';
+
+export type BaseIntegrationTask<Type extends IntegrationTaskType, Payload extends object> = {
+    type: Type;
+    payload: Payload;
+};
+
+export type IntegrationTaskImportSpaces = BaseIntegrationTask<
+    'import:spaces',
+    {
+        configQuery: string;
+        page?: string;
+        standaloneRef?: string;
+        /**
+         * The timestamp of the event that triggers the export.
+         *
+         * This is to help ensures that Git sync import and export operations are executed
+         * in the same order on GitBook and on the remote repository.
+         */
+        eventTimestamp?: Date;
+    }
+>;
+
+export type IntegrationTask = IntegrationTaskImportSpaces;