+++ integrations/gitlab/src/types.ts
@@ -10,8 +10,7 @@ export type GitLabRuntimeEnvironment = RuntimeEnvironment<{}, GitLabSpaceConfigu
 export type GitLabRuntimeContext = RuntimeContext<GitLabRuntimeEnvironment>;
 
 export type GitlabConfigureAction =
-    | { action: 'connect.gitlab'; withConnectGitLab: boolean }
-    | { action: 'save.token'; token: string }
+    | { action: 'save.token'; token: string; customInstanceUrl?: string }
     | { action: 'select.project'; project: string }
     | { action: 'select.branch'; branch: string }
     | { action: 'toggle.customTemplate'; withCustomTemplate: boolean }
@@ -70,8 +69,32 @@ export type GitlabConfigureProps = {
 
 export type GitlabConfigureState = Omit<SpaceInstallationConfiguration, 'project'> & {
     project?: string;
-    withConnectGitLab?: boolean;
     withCustomTemplate?: boolean;
     withCustomInstanceUrl?: boolean;
     commitMessagePreview?: string;
 };
+
+export type IntegrationTaskType = 'import:spaces';
+
+export type BaseIntegrationTask<Type extends IntegrationTaskType, Payload extends object> = {
+    type: Type;
+    payload: Payload;
+};
+
+export type IntegrationTaskImportSpaces = BaseIntegrationTask<
+    'import:spaces',
+    {
+        configQuery: string;
+        page?: string;
+        standaloneRef?: string;
+        /**
+         * The timestamp of the event that triggers the export.
+         *
+         * This is to help ensures that Git sync import and export operations are executed
+         * in the same order on GitBook and on the remote repository.
+         */
+        eventTimestamp?: Date;
+    }
+>;
+
+export type IntegrationTask = IntegrationTaskImportSpaces;